# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Replay já 2.0 — API
  version: 2.0.0-draft.1
  summary: Contrato entre o relay de gravação, o botão da quadra, o app do atleta e o painel do parceiro.
  description: |
    Contrato da API do Replay já 2.0.

    **Arquitetura**: a câmera IP da quadra **empurra RTMP** direto para um *relay* nosso
    (Lightsail, `sa-east-1`), que grava 24/7 em fMP4 com `ffmpeg -c copy` e corta o clipe a
    partir do índice. Não há computador de borda na arena. O botão é apenas um evento HTTP.

    | Superfície | Host | Autenticação |
    |---|---|---|
    | **Gatilho — botão físico** | `replayja.com.br/api/v1` | token no caminho (`/triggers/b/{buttonToken}`) |
    | **Gatilho — botão virtual** | `replayja.com.br/api/v1` | `userSession` |
    | **App do atleta** | `replayja.com.br/api/v1` | `userSession` (cookie HMAC próprio) |
    | **Painel do parceiro** | `replayja.com.br/api/v1` | `userSession` + `partner_admin` ativo |
    | **Relay → nuvem** | `replayja.com.br/api/relay` | `relayKey` (header `x-relay-key`) |
    | **Nuvem → relay** | `relay-1.replayja.com.br` | `relayToken` (HMAC no caminho) |

    **Quem manda em quem** (padrão herdado do relay v2 do Sentinela, em produção desde
    29/08/2026): o app **não escreve** no relay. O relay pergunta — lista de câmeras, fila de
    jobs de clipe — e a resposta é sempre uma *lista*, nunca um comando. A única chamada
    nuvem → relay é o `POST /jobs`, que apenas **acorda** o relay para buscar a fila mais cedo;
...

...
servers:
...
security:
  - userSession: []

# ---------------------------------------------------------------------------
paths:
...
  /triggers/b/{buttonToken}:
    post:
      tags: [Gatilhos]
      summary: Gatilho do botão físico (webhook)
      operationId: triggerFromButton
      security: []
      description: |
        Chamado **diretamente pelo botão** da quadra, um dispositivo Wi-Fi de bateria com
        webhook configurável (Shelly Button ou equivalente). Não há computador de borda: o
        botão associa no Wi-Fi da arena e faz um único `POST` HTTPS.

        **O contrato é deliberadamente mínimo**, porque o firmware do botão é de terceiro e
        muitos modelos não enviam corpo, cabeçalho custom nem assinatura:

        - O **segredo está no caminho** (`buttonToken`, 32 caracteres base62). É um segredo
          fraco por natureza — igual à chave RTMP — e a mitigação é o dano baixo (alguém
          dispara um clipe numa quadra), o rate limit e a revogação em um clique no painel.
        - **Corpo vazio é válido.** Nada é obrigatório.
        - **`Content-Type` é ignorado.**
        - A resposta é sempre `202`, mesmo em erro de aplicação — um botão de bateria não
          tem como tratar erro, e não queremos que ele repita.
        - Parâmetros opcionais de query, para os modelos que suportam template de URL:
          `?bat=<0-100>` (bateria), `?evt=<contador>` (contador do próprio botão, usado como
          chave de idempotência), `?rssi=<dBm>`.

        **O carimbo de tempo é a hora de chegada no servidor**, não a hora da cena. Duas
        latências separam uma da outra e as duas são descontadas no cálculo da janela:

        | Latência | Origem | Padrão | Como é obtida |
        |---|---|---|---|
        | `button.wakeLatencyMs` | botão sai do sono profundo, associa no Wi-Fi, resolve DNS, faz TLS | 1500 ms | medida na instalação, por câmera/botão |
        | `camera.originLagMs` | encoder da câmera + rede + buffer do ffmpeg no relay | 3000 ms | medida na instalação com relógio filmado |

        A janela pedida ao relay é
        `[t_chegada − wakeLatency − 24 s + originLag, t_chegada − wakeLatency + 1 s + originLag]`,
        e o relay corta uma janela **mais larga** (padrão `[−32 s, +6 s]`) para absorver erro
        das duas medidas e o alinhamento de segmento. O trecho entregue é recortado com
        precisão no passe de marca d'água.

        Rate limit: 1 gatilho por botão a cada 8 s (cooldown), 120/hora.
      parameters:
        - name: buttonToken
          in: path
          required: true
          schema: { type: string, pattern: '^[A-Za-z0-9]{24,48}$' }
        - name: bat
          in: query
          required: false
          description: Bateria em porcentagem, quando o botão sabe reportar.
          schema: { type: integer, minimum: 0, maximum: 100 }
        - name: evt
          in: query
          required: false
          description: Contador de eventos do próprio botão. Quando presente, vira chave de idempotência.
          schema: { type: integer, minimum: 0 }
        - name: rssi
          in: query
          required: false
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
          text/plain:
            schema: { type: string }
      responses:
        '202':
          description: |
            Sempre `202`, inclusive quando o gatilho é recusado (cooldown, câmera sem cobertura,
            botão revogado). O corpo diz o que aconteceu, para quem estiver depurando com curl;
            o botão ignora.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TriggerAck' }
        '404':
          description: Token inexistente. Único caso que não devolve 202 — evita manter token morto ativo para sempre.
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /auth/otp/start:
    post:
      tags: [Auth]
      summary: Envia o código OTP para o e-mail
      operationId: startOtp
      security: []
      description: |
        Único campo obrigatório: o e-mail. **Sem senha, sem nome, sem telefone, sem data de
        nascimento** — é o diferencial explícito frente ao cadastro pesado dos concorrentes
        (ver `docs/concorrentes.md`).

        O mesmo endpoint serve para cadastro e login: se o e-mail não existe, a conta é criada
        na verificação. O usuário nunca vê a distinção.

        `partnerSlug` é opcional e grava `app_user.first_partner_id` — atribuição de aquisição
        para o parceiro, mostrada no painel dele.

        Rate limit: 3 envios por e-mail a cada 15 min; 10 por IP/hora. A resposta é **sempre
        `202`**, mesmo para e-mail inválido ou bloqueado, para não permitir enumeração de contas.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                partnerSlug: { $ref: '#/components/schemas/Slug' }
                redirectTo:
                  type: string
                  format: uri
                  description: >-
                    Caminho relativo para onde voltar após o login
                    (ex. `/arena-calabouco/fut-segunda`). Validado contra o próprio domínio.
      responses:
        '202':
          description: Código enviado (resposta idêntica mesmo se o e-mail não puder receber)
          content:
            application/json:
              schema:
                type: object
                required: [expiresInSeconds]
                properties:
                  expiresInSeconds: { type: integer, examples: [600] }
                  resendAfterSeconds: { type: integer, examples: [60] }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    userSession:
      type: apiKey
      in: cookie
      name: session
      description: |
        Cookie de sessão assinado com HMAC-SHA256 (`<payload base64url>.<assinatura>`),
        `HttpOnly`, `Secure`, `SameSite=Lax`, validade de 400 dias com renovação deslizante
        a cada 7 dias de uso. Padrão portado do Sentinela (`lib/session.ts`).

        O payload é `{uid, email, exp}` e **não carrega autorização**: papel de admin de arena
        é consultado no banco na requisição de painel, para que remover um admin tenha efeito
        imediato em vez de esperar o cookie expirar.

        Não há tabela de sessão, nem provedor de identidade externo. Login por OTP de 6 dígitos
        (desafio também em cookie HMAC, de 10 min) ou por Google via OIDC.
    relayKey:
      type: apiKey
      in: header
      name: x-relay-key
      description: |
        Segredo por **relay** (não por câmera, não por arena), de 32 bytes, guardado em
        `/etc/replayja/rec.env` (modo 600, root) e com o SHA-256 no Postgres.

        Dá acesso a: a lista de câmeras daquele relay (incluindo as chaves RTMP), a fila de
        jobs daquele relay e as URLs de upload. **Não** dá acesso a nenhum dado de usuário.
        Rotação: mudar no painel, aplicar no relay, a chave antiga vale por 10 min.
    relayToken:
      type: apiKey
      in: header
      name: x-relay-key
      description: |
        Mesma credencial, na direção oposta: a API usa este cabeçalho ao chamar o relay
        (`POST /jobs`, `GET /stats`). Nas rotas `/t/{relayToken}/...` o token vai no **caminho**,
        porque ele é colocado num `<img src>` ou num link de download e não há como mandar
        cabeçalho — é um HMAC de escopo estreito (uma câmera, um intervalo) e validade curta,
        validado localmente pelo `auth-sidecar` do relay. Validar na nuvem viraria centenas de
        invocações por minuto.
    buttonToken:
      type: apiKey
      in: query
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        UUID gerado pelo cliente. Requisições repetidas com a mesma chave, no mesmo endpoint e
        dentro de 24h, devolvem a resposta original sem repetir o efeito. Ver
        `docs/api/README.md` §5.
      schema: { type: string, format: uuid }
  responses:
...
    TooManyRequests:
      description: Limite de taxa excedido
      headers:
        Retry-After:
          schema: { type: integer }
        RateLimit:
          description: RFC 9331 (`limit=…, remaining=…, reset=…`)
          schema: { type: string }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 9457 — Problem Details for HTTP APIs.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          examples: ['https://replayja.com.br/problems/clip-expired']
        title: { type: string, examples: ['Clipe expirado'] }
        status: { type: integer }
        detail:
          type: string
          description: Mensagem em pt-BR, pronta para exibir ao usuário.
        instance: { type: string }
        traceId: { type: string, description: 'Correlaciona com a linha em `app_error` e com o log da Vercel.' }
        errors:
          type: array
          description: Erros de validação campo a campo.
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
              message: { type: string }
...
    Slug:
      type: string
      pattern: '^[a-z0-9]+(-[a-z0-9]+)*$'
      minLength: 3
      maxLength: 40
      examples: ['arena-calabouco', 'fut-segunda']
...
    TriggerAck:
      type: object
      required: [accepted]
      properties:
        accepted: { type: boolean }
        triggerEventId: { type: string, format: uuid, nullable: true }
        clipId: { type: string, format: uuid, nullable: true }
        outcome:
          type: string
          enum: [accepted, rejected_cooldown, rejected_no_coverage, rejected_camera_unknown, rejected_button_revoked, rejected_relay_down]
        arrivalAt: { type: string, format: date-time }
        pressEstimatedAt:
          type: string
          format: date-time
          description: '`arrivalAt − button.wakeLatencyMs`. É o instante usado para montar a janela.'
        appliedLatencies:
          type: object
          properties:
            buttonWakeLatencyMs: { type: integer }
            cameraOriginLagMs: { type: integer }
        window:
          type: object
          properties:
            deliverFrom: { type: string, format: date-time }
            deliverTo: { type: string, format: date-time }
...